+++ cloudmanager/aws_fsx.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/fatih/structs"
-	"github.com/hashicorp/terraform/helper/schema"
 )
 
 // createAWSFSXDetails the users input for creating a FSX
@@ -26,7 +25,7 @@ type createAWSFSXDetails struct {
 	RouteTableIds          []string        `structs:"routeTableIds,omitempty"`
 	ThroughputCapacity     int             `structs:"throughputCapacity,omitempty"`
 	SecurityGroupIds       []string        `structs:"securityGroupIds,omitempty"`
-	AwsFSXTags             []fsxTags       `structs:"tags,omitempty"`
+	AwsFSXTags             []userTags      `structs:"tags,omitempty"`
 	TenantID               string
 }
 
@@ -73,18 +72,14 @@ type status struct {
 	Lifecycle string `json:"lifecycle"`
 }
 
-// expandFSXTags converts set to userTags struct
-func expandFSXTags(set *schema.Set) []fsxTags {
-	tags := []fsxTags{}
-
-	for _, v := range set.List() {
-		tag := v.(map[string]interface{})
-		fsxTag := fsxTags{}
-		fsxTag.TagKey = tag["tag_key"].(string)
-		fsxTag.TagValue = tag["tag_value"].(string)
-		tags = append(tags, fsxTag)
+// check if name tag exists
+func hasNameTag(tags []userTags) bool {
+	for _, v := range tags {
+		if v.TagKey == "name" {
+			return true
+		}
 	}
-	return tags
+	return false
 }
 
 func (c *Client) getAWSCredentialsID(name string, tenantID string) (string, error) {